Move to procps-ng.
[packaging] Update to procps-ng 3.3.15. Contributes to JB#44378
Clean up a bunch of patches for things that have been fixed in
upstream and removed from Fedora as well.

Clean up the spec file. Some lines are borrowed from Fedora. Include
symlinks for binaries that used to be in /bin and /sbin. Add devel
subpackage for pkgconfig and library symlink.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>